Move makeDestination helper into nitro-protocol
