+++ packages/nitro-protocol/src/index.ts
@@ -87,4 +87,4 @@ import * as Transactions from './transactions';
 export {Signatures, Transactions};
 
 // types
-export {Uint256, Bytes32, Destination} from './contract/types';
+export {Uint256, Bytes32, Destination, toUint256, makeDestination} from './contract/types';
